+++ CHANGELOG.md
@@ -14,6 +14,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     * Fix `flatpick` for readonly property.
 * Now settings for `detail-interation` services are applying correctly on closing detail's edit form.
 
+### Removed
+* `flexberry-simpleolv` component:
+    * Remove `_attributeChanged` observers.
+
 ## [0.7.0-beta.26] - 2016-10-22
 ### Changed
 * Updated dependency on `ember-flexberry-data` addon to v0.6.2-beta.30 version.